This is a basic and original set of functions about the linked list.
ll_born: It can produce the first node of a linked list.

cll_born: It can produce the first node of a circular linked list.

ll_ins: It can insert a node into a (circular) linked list next to the head node if there is. I name that inserting operation neck-insertion personally. When there is no head node, which means that the node to be inserted is the first node of a (circular) linked list, it will call ll_born or cll_born to produce a node. So, it may looks like "1->7->6->5->4->3->2" after inserting 1 to 7 into a NULL ordinary linked list.

ll_init: It can initialize a (circular) linked list with some nodes whose key pointers are NULL.

ll_del: It can delete a node next to the head node if the number of nodes is greater than 1. When the number of nodes is less than or equal to 1, it can delete the very node if existing and return NULL. Please notice that, this function can be applied not only on the head node but also any node of a (circular) linked list except the last one of an ordinary linked list. Deleting the last node of an ordinary linked list will produce a wild pointer.

ll_burn: It can burn and destory a (circular) linked list clean. It can be applied on any node of a circular linked list leading to the same result. But for an ordinary linked list, I suggest that do apply it on the head node rather than the others forever to avoid the wild pointer efficiently.