Add a message queue and a task for the ISR of CAN receiving.
What I want to do is to reduce time consuming of the ISR. And it's not a good idea to use malloc in an ISR.

After adding a task, modify priorities of tasks.
The task_can's  priority is 100,
The task_gps's  priority is 110,
The task_accs's priority is 120,
The task eots's priority is 130,
The task mls's  priority is 140,
The taks_sync's priority is 150.